refactor: extract closeChannel logic into objective
